requirements.txt
python-dotenv
SQLAlchemy
pandas
pymysql
cryptography
requests